Fix: Print for student report not showing images of image questions.
This problem was intermittant, and the result of images not being (consistently) loaded on
the student page shortly after the print button was pressed.

After taking a deep dive into using image onLoad callbacks &etc,

I decided it was simpler to hide the student view (instead of not rendering), so the images would
be pre-loaded for printing.  There may be performance issues with this down the road.

NB: There is an issue with firefox not firing media query events after the print dialog is closed,
so that we remain on the students report page.

[#115800099]
https://www.pivotaltracker.com/story/show/115800099