Feature:  Save & Load rubric (content) using report API.
We now 'Freeze` rubric after it has been loaded from a URL.  This is to prevent rubric from changing after a teacher has
graded some or all of their students.

General process:

1. When report data is first loaded from the server, we load rubric for activity feedback, if it exists in serialized form.
2. After the report data has all been loaded, we check for missing activityFeedback rubric for which we have a URL.
3. If the URL exists, but we haven't loaded a copy of the rubric yet, we load it.
4. We update any activities that don't have a serialized version of the rubric JSON using the portal `enableActivityFeedback` api

Main changes:
1. The `rubricReducer` now handles `RECEIVE_DATA` actions.  It looks for serialized rubric. If it finds some it puts it into the state.
2. The 'addRubric` action now dispatches `enableActivityFeedback` actions for activities that are missing serialized rubric.

PT Story:
If a teacher uses a rubric to give feedback, it is copied and locked for them for that assignment, so that it does not change if the author changes it after that point.

[#155368181]

https://www.pivotaltracker.com/story/show/155368181