Fix:  Immutable wrap / unwrap of Rubric via both load paths.
The rubric could be loaded two ways.  Sometimes it was being wrapped as an Immutable Map, sometimes it was a raw object.
The solution was to to force saving it as a Map in the store, and call toJS() in view.

PT Story:
If a teacher uses a rubric to give feedback, it is copied and locked for them for that assignment, so that it does not change if the author changes it after that point.

[#155368181]

https://www.pivotaltracker.com/story/show/155368181