Feature: Default max score is 10 instead of 0.
Also: No scores less than 1.

Story:
[#130389079]
https://www.pivotaltracker.com/story/show/130389079

Right now when you add scoring as a feedback option the max score is "0" which is confusing and potentially problematic if we do any calculations that might end up with a division by zero error. Let's make it 10, and actually prevent anything below one.