Feature: Rubric based Automatic Scores
Story:
[#155408233]

If there is a rubric, with scores, and the teacher uses the rubric, the default scoring method should be "Automatic Scoring using rubric scores".

If the rubric doesn't exist, or if it doesn't include scores, do not show the second option "Automatic Scoring using rubric scores" -- don't even gray it out.

https://www.pivotaltracker.com/story/show/155408233