Fixes crashing when displaying answer-table without selected question.
The app was crashing when the question detail box is being closed while showing student work.

It seems the question prop becomes null somewhere in the viewheirarchy when the close button is pressed.

A simple null check on in the  answer-table rendering stops the bug.

I added tests for the container that fail without the null check.

[#164583438]

https://www.pivotaltracker.com/story/show/164583438