Hot-fix: CSS (boostrap modal) regression in CompareView
bootstrap and react-bootstrap were updated along with React 16.
Animation interfaces must have broken, because faded components never
un-fade.

This fix is a little bit of a hack because it doesn't restore animation, it just makes
sets faded bootstrap components to opaque instead of transparent.

https://www.pivotaltracker.com/story/show/166907544
[#166907544] fix-compare-view

Compare view no longer works.

This is a regression from v1.9.0 to v1.10.0, probably a result of updating React to 16.