[#166240880] FireStore: Save assignment level report settings
We save the teachers settings for the offering at this FireStore path:
/sources/${sourceId}/user_settings/${platformUserId}/offering/${contextId}

We dispatch all user settings updates to FireStore via the middleware API.
The RECEIVE_USER_SETTINGS action is the one which changes the active store.

Even when the network is disabled (wifi disconnected) the FireStore API dispatch method
works, and the RECEIVE_USER_SETTINGS is reliably called with valid parameters.

The FireStore rules are setup to allow unfettered access to:
`/sources/fake.portal/user_settings/1/offering/class123`

This allows us to test the report settings with fake portal data, without a valid JWT.

https://www.pivotaltracker.com/story/show/166240880