Fix correct answer lookup and add some tests
I'm not really sure the tests are worth it. They at least illustrate what state is used
to compute the correct answer. But I think they will be hard to maintain. The number
or references between objects based on ids makes it really easy to have a mismatch
and then the tests fails for a very hard to find reason.

Also the tests are based on looking at what the code needs just to get it to return the
right value. They don't really confirm that the data from the servers will result in the
right value.

A more useful test would be one that runs agains the data in the js/data and checks
the result, but I don't think we have an example of that.