disable network on the demo report, handle uninitialized user settings
disabling the network makes the demo report behave more consistently with the what the cypress tests see
this change also required updating the tests now that there wasn't some pre-existing teacher settings

the uninitialized user settings seems to only happen when the network is disabled. I didn't track it down,
but I suspect there is some initializer that is setting some initial user settings. And that is why the
user settings always had a visibility_filter when the network is working properly.