add automated test of feedback scores
This test currently fails.
Also this is only testing that the teacher sees the correct score in the disabled score box.
More important would be to check that the student sees the score in their feedback.
However we don't yet have a good pattern to mock the student feedback data now that
the feedback has moved to firestore.