Better 🐞 fix  for Activity scrambled in dashboard
A better fix for 🐞 Activity order scrambled in teacher dashboard
[#170757052]

The previous PR sorted the activities (from activityTree) by the position attributes.

That sorting is not required if we use the `getSequenceTree()` Selector which is careful to
map deep references to children by `id` from the shallow normalized representation of the children array.

`report-tree_spec.js` was modified to use keys instead of numeric indexes.

The activity map in the spec was modified so that the keys were out of sort order.

[#170757052]

https://www.pivotaltracker.com/story/show/170757052