Show standalone iframe for `iframeQuestionId` url param
If the url includes the param `iframeQuestionId`, we will switch to a
new base component `<IframeStandaloneApp>`.

The fetching of the portal data will still proceed as usual, but the
IframeStandaloneApp component will select the question and answer
specified by the new param and the `studentId` param. If no answer
is found, an error is displayed.