Add link in main report to open iframes in standalone mode
This appends the link "Open in new tab" to any iframe answers that
currently are shown embedded in the report. It does not affect iframes
that already open in a new tab, such as CODAP

(Note: the standalone-iframes url will also work fine for CODAP-style
models, as it will use the reporting url in the standalone iframe.
However, it seems unnecessary to add an additional iframe wrapping to
these links. This may be useful for researcher reports, however.)

This also makes slight changes to the links UI, to allow the "Open in
new tab" link to be always visible.