View, Data, Feedback, Dashboard models -> StaticallyTypedRecord
This is the first step of adding types to our Redux models. This supports
key existence and type checking for code such as `viewState.get("type")`
or `viewState.set("type", "dashboard").

This type checking won't yet be pushed out to the rest of the application
until the root state is typed.

This converts the Maps to Records, which supports typing better, but the
initial `Record.Factory` that is created by `RecordFactory` then creates
a `Record` class, which needs to be called with `new`.

It is kind of odd that `INITIAL_VIEW_STATE` is a class, but this naming
will become a little more logical once we have to introduce classes that
extend this initial state.

Note that the original `record-factory.ts` code is from
https://blog.mgechev.com/2018/01/18/react-typescript-redux-immutable/
and had a different `RecordFactory` function:

```
export const RecordFactory = <T>(seed: T): StaticallyTypedRecord<T> => {
  return (Record(seed) as any) as StaticallyTypedRecord<T>;
};
```

This did not work as the returned Records were of type `IViewState`
instead of `StaticallyTypedRecord<IViewState>`, and so did not support
the Record functions such as `get`, `set`, etc.