Add types to Report reducer
This uncovered a large number of properties of the Report model that
were not apparent in the original initial data, as well as types
missing in the `transform-json-response` code.

This also introduces using a new class, `ReportState`, which at the
moment simply extends `INITIAL_REPORT_STATE`, but allows us to use it
as a type rather than something cumbersome like
`type ReportStateType = typeof INITIAL_REPORT_STATE`. It also leads
towards the classes we'll need to let `getIn` correctly type-check
deeper properties.

The same class pattern is added to the other reducers.