Use RootState as type where possible; update types
This pushes out the types to the rest of the application where possible
(probably several places missed), and most particularly in
`actions/index.ts`.

By using `RootState`, this exposes an issue with the up-til-now
implementation, which was that the `getIn([x, y])` function was not
correctly returning types, as the nested models don't correctly
enumerate their properties. In order for this to work correctly, the
classes such as `ReportState` etc need to explicitly implement the
interfaces such as `IReportState`, which do enumerate those properties.

However, saying we implement `IReportState` requires us to actually add
the props to the class definition, and would also normally require us to
add values to each prop, or we get the error `Property 'type' has no
initializer and is not definitely assigned in the constructor.ts(2564)`

Putting in all the default values a second time seemed like a recipe for
confusion and potential bugs down the road, so adding
`"strictPropertyInitialization": false,` to the tsconfig seemed  like
the route of least harm.

This isn't an ideal, as first of all the classes *don't* really implement
those interfaces, they return functions that create Records implementing
those interfaces, and second it requires duplicating the list of
properties. However, it allows us to actually use all the strong typings
defined in the stores, and is also the method described in
https://blog.mgechev.com/2018/01/18/react-typescript-redux-immutable/