fix the jest tests, and better support jwt requests
The portal JWT api currently only supports a resource_link_id if the current user
is a student with that resource_link_id (a learner with an offering id that matches the resource_link id)
teacher users cannot have this resource_link_id param.

This code improves things so the resource_link_id param is only added to the jwt
request if the report url has a studentId. However what we really need is to know if the
current user is a student or teacher.