Fix error when displaying fetch errors
This also should include full coverage of the DataFetchError component

The previous implementation was actually broken, because it was assuming the
error.body property was a string. When the error object is actually the response
from a fetch, its body is a ReadableStream. React can't handle rendering a
ReadableStream. So the DataFetchError component needed to be expanded
so it reads the text out of the stream in order to render it.