fix errors in cypress after updating
"not.be.visible" fails in cypress 6 if the element doesn't exist

`within` yields the same subject it was given from the previous command. This behavior is
not supposed to have changed from cypress 5 to 6. Many of the helper functions seemed
to be assuming within would return the last value selected inside of the within block.
I don't understand why tests using these helper functions passed before.