Add researcher support
This uses the new portal researcher JWT support,
and the new report-service researcher support.
It now passes the studentId param into the queries.
This is needed so the firestore rules can only allow access to a specific student
based on the JWT issued by the portal.

This requires more work:
- it needs tests
- I need to understand and clarify some comments I left before