update tests to cover the sharedWith case more clearly
Besides the test updates this includes a few runtime changes:
- renames fake-user-type to fakeUserType for consistency.
- adds a fakePlatformUserId parameter so the studentId and platformUserId in the fake
portal data can be matched or not more explicitly.
- adds mapping of shared_with param so feedback requests don't crash the report when
the studentId doesn't match the platformId

The actual coverage of the sharedWith case is hacky because I don't have time to
do it better. It takes advantage of the fact that feedback is still requested from firestore
even when fake answers are being used.